+++ src/main/PDF/Services/CrudServices/UploadPDFService.java
@@ -11,19 +11,8 @@ import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.gridfs.GridFSBucket;
 import com.mongodb.client.gridfs.GridFSBuckets;
 import com.mongodb.client.gridfs.model.GridFSUploadOptions;
-import org.apache.commons.io.output.ByteArrayOutputStream;
-import org.apache.pdfbox.pdmodel.PDDocument;
-import org.apache.pdfbox.pdmodel.PDPage;
-import org.apache.pdfbox.pdmodel.PDPageContentStream;
-import org.apache.pdfbox.pdmodel.graphics.image.LosslessFactory;
-import org.apache.pdfbox.pdmodel.graphics.image.PDImageXObject;
 import org.bson.Document;
 
-import javax.imageio.ImageIO;
-import java.awt.geom.AffineTransform;
-import java.awt.image.AffineTransformOp;
-import java.awt.image.BufferedImage;
-import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.security.GeneralSecurityException;
@@ -75,11 +64,14 @@ public class UploadPDFService implements Service {
       return PdfMessage.INVALID_PDF;
     } else {
       if (fileContentType.startsWith("image")) {
-        try {
-          fileStream = convertImageToPDF(fileStream);
-          filename = filename.substring(0, filename.lastIndexOf(".")) + ".pdf";
-        } catch (IOException exception) {
+        ImageToPDFService imageToPDFService = new ImageToPDFService(fileStream);
+        Message response = imageToPDFService.executeAndGetResponse();
+
+        if (response == PdfMessage.INVALID_PDF) {
           return PdfMessage.INVALID_PDF;
+        } else {
+          fileStream = imageToPDFService.fileStream;
+          filename = filename.substring(0, filename.lastIndexOf(".")) + ".pdf";
         }
       }
       if ((pdfType == PDFType.COMPLETED_APPLICATION
@@ -134,60 +126,5 @@ public class UploadPDFService implements Service {
     }
     gridBucket.uploadFromStream(filename, inputStream, options);
     return PdfMessage.SUCCESS;
-  }
-
-  public InputStream convertImageToPDF(InputStream fileStream) throws IOException {
-    final int PDFHEIGHT = 792;
-    final int PDFWIDTH = 612;
-
-    PDDocument document = new PDDocument();
-    PDPage page = new PDPage();
-    document.addPage(page);
-    BufferedImage bimg = ImageIO.read(fileStream);
-    float imageWidth = bimg.getWidth();
-    float imageHeight = bimg.getHeight();
-
-    if (imageWidth > imageHeight) {
-      bimg = rotateImage(bimg);
-      float tmp = imageWidth;
-      imageWidth = imageHeight;
-      imageHeight = tmp;
-    }
-
-    float finalWidth, finalHeight;
-    if (imageWidth / PDFWIDTH > imageHeight / PDFHEIGHT) {
-      finalWidth = PDFWIDTH;
-      finalHeight = imageHeight * PDFWIDTH / imageWidth;
-    } else {
-      finalHeight = PDFHEIGHT;
-      finalWidth = imageWidth * PDFHEIGHT / imageHeight;
-    }
-
-    PDImageXObject imageXObject = LosslessFactory.createFromImage(document, bimg);
-    PDPageContentStream contentStream = new PDPageContentStream(document, page);
-    contentStream.drawImage(imageXObject, 0, 0, finalWidth, finalHeight);
-    contentStream.close();
-    ByteArrayOutputStream out = new ByteArrayOutputStream();
-    document.save(out);
-    document.close();
-    ByteArrayInputStream inputStream = new ByteArrayInputStream(out.toByteArray());
-    return inputStream;
-  }
-
-  public BufferedImage rotateImage(BufferedImage bimg) {
-    // https://blog.idrsolutions.com/2019/05/image-rotation-in-java/
-    final double rads = Math.toRadians(90);
-    final double sin = Math.abs(Math.sin(rads));
-    final double cos = Math.abs(Math.cos(rads));
-    final int w = (int) Math.floor(bimg.getWidth() * cos + bimg.getHeight() * sin);
-    final int h = (int) Math.floor(bimg.getHeight() * cos + bimg.getWidth() * sin);
-    final BufferedImage rotatedImage = new BufferedImage(w, h, bimg.getType());
-    final AffineTransform at = new AffineTransform();
-    at.translate(w / 2, h / 2);
-    at.rotate(rads, 0, 0);
-    at.translate(-bimg.getWidth() / 2, -bimg.getHeight() / 2);
-    final AffineTransformOp rotateOp = new AffineTransformOp(at, AffineTransformOp.TYPE_BILINEAR);
-    rotateOp.filter(bimg, rotatedImage);
-    return rotatedImage;
   }
 }
